Build libskia.so from source.
The library /system/vendor/lib/libqc-opt.so relies on the stock libskia.so
because of some symbol called "co_node." So we edit libqc-opt.so's linked
libraries to change libskia to libLGsk.so (must be same size) and rename
LG's libskia.so to libLGsk.so, and build libskia.so from source. Now
everybody is happy.

Change-Id: I8613052df6338d7c0e7b3c41bea6cd498b35a6e1